% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCalc.R
\name{perimetre_carre}
\alias{perimetre_carre}
\title{Calcul du périmètre d'un carré}
\usage{
perimetre_carre(cote)
}
\arguments{
\item{cote}{La longueur d'un côté du carré (doit être positive).}
}
\value{
Le périmètre du carré.
}
\description{
Calcul du périmètre d'un carré
}
\examples{
perimetre_carre(5)
}
