% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCalc.R
\name{perimetre_rectangle}
\alias{perimetre_rectangle}
\title{Calcul du périmètre d'un rectangle}
\usage{
perimetre_rectangle(longueur, largeur)
}
\arguments{
\item{longueur}{La longueur du rectangle.}

\item{largeur}{La largeur du rectangle.}
}
\value{
Le périmètre du rectangle.
}
\description{
Calcul du périmètre d'un rectangle
}
\examples{
perimetre_rectangle(5, 3)
}
