% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCalc.R
\name{surface_carre}
\alias{surface_carre}
\title{Calcul de la surface d'un carré}
\usage{
surface_carre(cote)
}
\arguments{
\item{cote}{La longueur d'un côté du carré.}
}
\value{
La surface du carré.
}
\description{
Calcul de la surface d'un carré
}
\examples{
surface_carre(5)
}
