% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCalc.R
\name{surface_rectangle}
\alias{surface_rectangle}
\title{Calcul de la surface d'un rectangle}
\usage{
surface_rectangle(longueur, largeur)
}
\arguments{
\item{longueur}{La longueur du rectangle.}

\item{largeur}{La largeur du rectangle.}
}
\value{
La surface du rectangle.
}
\description{
Calcul de la surface d'un rectangle
}
\examples{
surface_rectangle(5, 3)
}
